Fix toc (#58)
* TOC: use kebab case instead of camel case, update test

* update python script to use kebab case

* update docs

* resolve conflict

* update toc

* fix out of range

* handle non existing old files

* update explanation

* add render script

* fix toKebab, toSnake, etc

* update command

* better edge case handling

* start using built-in

* Better docs

* add progress

+++ toc/tocitem.go
@@ -10,7 +10,7 @@ import (
 type TocItem struct {
 	Level           int
 	Title           string
-	CamelCaseTitle  string
+	KebabCaseTitle  string
 	OldFileLocation string
 	NewFileLocation string
 	Children        TocItems
@@ -35,12 +35,11 @@ func (tocItem *TocItem) RenderNewContent() (err error) {
 		if err := tocItem.RenderNewContentToNewFile(tocHeader, tocSubtopic); err != nil {
 			return err
 		}
+		return tocItem.Children.RenderNewContent()
 	}
 	// old file location is defined
-	if tocItem.OldFileLocation != "" {
-		if err := tocItem.RenderNewContentFromOldFile(tocHeader, tocSubtopic); err != nil {
-			return err
-		}
+	if err := tocItem.RenderNewContentFromOldFile(tocHeader, tocSubtopic); err != nil {
+		return err
 	}
 	// also render children
 	return tocItem.Children.RenderNewContent()
@@ -58,7 +57,10 @@ func (tocItem *TocItem) RenderNewContentFromOldFile(tocHeader, tocSubtopic strin
 		return err
 	}
 	if tocItem.OldFileLocation != tocItem.NewFileLocation {
-		return os.Remove(tocItem.OldFileLocation)
+		oldFileExist, oldFileExistErr := util.File.IsExist(tocItem.OldFileLocation)
+		if oldFileExist && oldFileExistErr == nil {
+			return os.Remove(tocItem.OldFileLocation)
+		}
 	}
 	return nil
 }
@@ -146,7 +148,7 @@ func (tocItem *TocItem) SetNewFileLocation() {
 	pathList := []string{tocItem.getNewFileName()}
 	parent := tocItem.Parent
 	for parent != nil {
-		pathList = append([]string{parent.CamelCaseTitle}, pathList...)
+		pathList = append([]string{parent.KebabCaseTitle}, pathList...)
 		parent = parent.Parent
 	}
 	pathList = append([]string{tocDirPath}, pathList...)
@@ -155,9 +157,9 @@ func (tocItem *TocItem) SetNewFileLocation() {
 
 func (tocItem *TocItem) getNewFileName() (fileName string) {
 	if len(tocItem.Children) == 0 {
-		return fmt.Sprintf("%s.md", tocItem.CamelCaseTitle)
+		return fmt.Sprintf("%s.md", tocItem.KebabCaseTitle)
 	}
-	return filepath.Join(tocItem.CamelCaseTitle, "README.md")
+	return filepath.Join(tocItem.KebabCaseTitle, "README.md")
 }
 
 func NewTocItem(toc *Toc, parent *TocItem, level int, title, oldFileLocation string) (tocItem *TocItem) {
@@ -169,7 +171,7 @@ func NewTocItem(toc *Toc, parent *TocItem, level int, title, oldFileLocation str
 	tocItem = &TocItem{
 		Level:           level,
 		Title:           title,
-		CamelCaseTitle:  toc.Util.Str.ToCamel(title),
+		KebabCaseTitle:  toc.Util.Str.ToKebab(title),
 		OldFileLocation: oldFileLocation,
 		Parent:          parent,
 		Toc:             toc,
